Replace GitHub cloud sync action
This changeset replaces the repo's outdated "s3 cloud sync" action
with the hubverse-aws-upload action from the Hubverse's official
actions repository.

This also adds a cloud section to admin.json, for review of
the bucket name. We'll leave the cloud disabled for now, so the
new github action should be a no-op.